edit

+++ 0x00-python_variable_annotations/0-add.py
@@ -2,6 +2,6 @@
 """Variable annotation introduction alx"""
 
 
-def add(a: int, b: int) -> int:
+def add(a: float, b: float) -> int:
     """ return the sum of a and b """
     return a + b
